Library: Fix First Load Bug
In the Library project add a fix for a bug that occurred when users loaded the page for the first time and no localstorage was found.

Now when loading the user data, if no localstorage key is found it generates an empty array - otherwise it loads the array saved in localstorage.